refactor: clean code (delete comments, console.log(), ...)
